Move feature-specific tests into their own folder
