+++ e2e/playwright/feature/dark-mode.spec.ts
@@ -4,9 +4,9 @@ import {
   TestingLibraryFixtures,
 } from '@playwright-testing-library/test/fixture'
 
-import { LoginPage } from '../../models/Login'
-import { resetDb } from '../../cms/database/seed'
-import { seedDB } from '../database/seedMongo'
+import { LoginPage } from '../models/Login'
+import { resetDb } from '../cms/database/seed'
+import { seedDB } from '../portal-client/database/seedMongo'
 
 type CustomFixtures = {
   loginPage: LoginPage
@@ -26,10 +26,13 @@ test.beforeAll(async () => {
   await seedDB()
 })
 
-describe('MySpace', () => {
-  test('can login and see MySpace', async ({ page, loginPage }) => {
+describe('Toggle light/dark mode', () => {
+  test('can toggle light/dark mode', async ({ page, loginPage }) => {
     await loginPage.login('user1', 'user1pass')
     await expect(page.locator('text=WELCOME, BERNIE')).toBeVisible()
-    await expect(page.locator('text=Example Collection')).toBeVisible()
+    await expect(page.locator('text=DARK MODE')).toBeVisible()
+
+    await page.locator('[data-testid="theme-toggle"]').click()
+    await expect(page.locator('text=LIGHT MODE')).toBeVisible()
   })
 })
